Moved CTextInput to a separate file. No changes in logic
