+++ client/globalLobby/GlobalLobbyWidget.cpp
@@ -21,11 +21,11 @@
 #include "../gui/CGuiHandler.h"
 #include "../gui/WindowHandler.h"
 #include "../widgets/Buttons.h"
+#include "../widgets/CTextInput.h"
 #include "../widgets/GraphicalPrimitiveCanvas.h"
 #include "../widgets/Images.h"
 #include "../widgets/MiscWidgets.h"
 #include "../widgets/ObjectLists.h"
-#include "../widgets/TextControls.h"
 
 #include "../../lib/CConfigHandler.h"
 #include "../../lib/Languages.h"
