+++ client/lobby/CSavingScreen.cpp
@@ -17,7 +17,7 @@
 #include "../gui/CGuiHandler.h"
 #include "../gui/Shortcut.h"
 #include "../widgets/Buttons.h"
-#include "../widgets/TextControls.h"
+#include "../widgets/CTextInput.h"
 
 #include "../../CCallback.h"
 #include "../../lib/CConfigHandler.h"
